Keep stub servlet as internal class.
