+++ vaadin-spring/src/test/java/com/vaadin/flow/spring/SpringClassesSerializableTest.java
@@ -66,7 +66,7 @@ public class SpringClassesSerializableTest extends ClassesSerializableTest {
                 "com\\.vaadin\\.flow\\.spring\\.VaadinServletContextInitializer\\$ErrorParameterServletContextListener",
                 "com\\.vaadin\\.flow\\.spring\\.VaadinServletContextInitializer\\$DevModeServletContextListener",
                 "com\\.vaadin\\.flow\\.spring\\.VaadinServletContextInitializer\\$WebComponentServletContextListener",
-                "com\\.vaadin\\.flow\\.spring\\.StubServletConfig",
+                "com\\.vaadin\\.flow\\.spring\\.VaadinServletContextInitializer\\$SpringStubServletConfig",
                 "com\\.vaadin\\.flow\\.spring\\.VaadinMVCWebAppInitializer",
                 "com\\.vaadin\\.flow\\.spring\\.RootMappedCondition",
                 "com\\.vaadin\\.flow\\.spring\\.VaadinWebsocketEndpointExporter",
